+++ packages/twopoint5d/CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to [@spearwolf/twopoint5d](https://github.com/spearwolf/twop
 The format is loosely based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.17.0] - 2025-11-25
 
 - add `frameRate` (fps) option as alternative to `duration` in `FrameBasedAnimations`
   - the `add()` method now accepts either a `duration` number or an `AnimationTimingOptions` object with `frameRate` or `duration`